+++ sdk/resourcemanager/subscription/armsubscription/zz_generated_subscriptions_client.go
@@ -36,7 +36,7 @@ func NewSubscriptionsClient(credential azcore.TokenCredential, options *arm.Clie
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -53,6 +53,7 @@ func NewSubscriptionsClient(credential azcore.TokenCredential, options *arm.Clie
 
 // Get - Gets details about a specified subscription.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2016-06-01
 // subscriptionID - The ID of the target subscription.
 // options - SubscriptionsClientGetOptions contains the optional parameters for the SubscriptionsClient.Get method.
 func (client *SubscriptionsClient) Get(ctx context.Context, subscriptionID string, options *SubscriptionsClientGetOptions) (SubscriptionsClientGetResponse, error) {
@@ -84,7 +85,7 @@ func (client *SubscriptionsClient) getCreateRequest(ctx context.Context, subscri
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2016-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -99,9 +100,10 @@ func (client *SubscriptionsClient) getHandleResponse(resp *http.Response) (Subsc
 
 // NewListPager - Gets all subscriptions for a tenant.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2016-06-01
 // options - SubscriptionsClientListOptions contains the optional parameters for the SubscriptionsClient.List method.
 func (client *SubscriptionsClient) NewListPager(options *SubscriptionsClientListOptions) *runtime.Pager[SubscriptionsClientListResponse] {
-	return runtime.NewPager(runtime.PageProcessor[SubscriptionsClientListResponse]{
+	return runtime.NewPager(runtime.PagingHandler[SubscriptionsClientListResponse]{
 		More: func(page SubscriptionsClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
@@ -138,7 +140,7 @@ func (client *SubscriptionsClient) listCreateRequest(ctx context.Context, option
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2016-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -154,11 +156,12 @@ func (client *SubscriptionsClient) listHandleResponse(resp *http.Response) (Subs
 // NewListLocationsPager - This operation provides all the locations that are available for resource providers; however, each
 // resource provider may support a subset of this list.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2016-06-01
 // subscriptionID - The ID of the target subscription.
 // options - SubscriptionsClientListLocationsOptions contains the optional parameters for the SubscriptionsClient.ListLocations
 // method.
 func (client *SubscriptionsClient) NewListLocationsPager(subscriptionID string, options *SubscriptionsClientListLocationsOptions) *runtime.Pager[SubscriptionsClientListLocationsResponse] {
-	return runtime.NewPager(runtime.PageProcessor[SubscriptionsClientListLocationsResponse]{
+	return runtime.NewPager(runtime.PagingHandler[SubscriptionsClientListLocationsResponse]{
 		More: func(page SubscriptionsClientListLocationsResponse) bool {
 			return false
 		},
@@ -193,7 +196,7 @@ func (client *SubscriptionsClient) listLocationsCreateRequest(ctx context.Contex
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2016-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }